platforms: remove unnecessary int64 cast in splitUnixTime

+++ common_64.go
@@ -8,5 +8,5 @@ import "time"
 
 // splits unix time given in nanoseconds into secs and nsecs parts
 func splitUnixTime(utime int64) (int64, int64) {
-	return int64(utime) / int64(time.Second), int64(utime) % int64(time.Second)
+	return utime / int64(time.Second), utime % int64(time.Second)
 }
